+++ CHANGELOG.md
@@ -16,6 +16,7 @@
     - `GridArray::repeat()` for constructing arrays with uniform contents.
     - `GridArray::from_element()` for constructing single-element arrays.
     - `GridRotation::ALL_BUT_REFLECTIONS`, as `GridRotation::ALL` but excluding reflections.
+    - `SpaceBuilder::filled_with()` allows specifying the initial block in a new `Space`.
     - `SpaceTransaction::draw_target()` allows 2.5D drawing into a transaction in the same way `Space::draw_target()` works on `&mut Space`.
     - `StandardCameras::world_space()`, so that clients don't need to consult the `Character` each frame.
     - `Universe::get_any()` allows looking up universe members without knowing their type.